+++ AddOns/BetterBlizzFrames/midnight/modules/healthbarColor.lua
@@ -553,6 +553,7 @@ local function HookPowerBarColors()
 
     if customColorsRaidFrames and not BBF.powerColorsRaidFramesHooked then
         hooksecurefunc("CompactUnitFrame_UpdatePowerColor", function(frame)
+            if issecretvalue(frame) then return end
             if not frame or not frame.unit or frame.unit:find("nameplate") or frame:IsForbidden() then return end
 
             local _, powerToken = UnitPowerType(frame.unit)
@@ -978,6 +979,7 @@ function BBF.HookHealthbarColors()
 
         if (rpNamesHealthbarColor and TRP3_API) or customHealthbarColors then
             local function UpdateHealthColorUnified(frame)
+                if issecretvalue(frame) then return end
                 if not frame or not frame.unit or frame.unit:find("nameplate") or frame:IsForbidden() then return end
 
                 if TRP3_API and rpNamesHealthbarColor then