+++ AddOns/BetterBlizzFrames/midnight/modules/partyHighlight.lua
@@ -16,6 +16,7 @@ local currentAtlas = "RaidFrame-AgroFrame"
 local currentColor = {0, 1, 0, 1}
 
 local function ApplyHighlight(frame)
+    if issecretvalue(frame) then return end
     if not frame or frame:IsForbidden() then return end
     if not frame.selectionHighlight then return end
     if frame.unit:find("nameplate") then return end