+++ AddOns/BetterBlizzFrames/CHANGELOG.md
@@ -1,3 +1,17 @@
+# BetterBlizzFrames 1.9.3i
+## Midnight
+### Bugfix
+- Add some secret checks to avoid potential errors.
+
+# BetterBlizzFrames 1.9.3h
+## Midnight
+### Bugfix
+- Fix healthbar sticking out a bit from frame on minus mobs for classic frames setting.
+- Fix castbar color logic if classic castbars were enabled but not classic frames.
+- Fix "DarkMode: Tooltip" covering gems on tooltip with dark overlay.
+- Fix "No Portrait: Pixel Border" settings pixel border on minus mobs' manabar being out of position.
+- Fix some player name changes not going through on login like RP names.
+
 # BetterBlizzFrames 1.9.3g
 ## Midnight
 ### Bugfix